+++ src/Sentry/BindableSentryOptions.cs
@@ -53,6 +53,15 @@ internal partial class BindableSentryOptions
     public bool? EnableSpotlight { get; set; }
     public string? SpotlightUrl { get; set; }
 
+    [Experimental(Infrastructure.DiagnosticId.ExperimentalFeature)]
+    public BindableSentryExperimentalOptions Experimental { get; set; } = new();
+
+    [Experimental(Infrastructure.DiagnosticId.ExperimentalFeature)]
+    internal sealed class BindableSentryExperimentalOptions
+    {
+        public bool? EnableLogs { get; set; }
+    }
+
     public void ApplyTo(SentryOptions options)
     {
         options.IsGlobalModeEnabled = IsGlobalModeEnabled ?? options.IsGlobalModeEnabled;
@@ -99,6 +108,8 @@ internal partial class BindableSentryOptions
         options.JsonPreserveReferences = JsonPreserveReferences ?? options.JsonPreserveReferences;
         options.EnableSpotlight = EnableSpotlight ?? options.EnableSpotlight;
         options.SpotlightUrl = SpotlightUrl ?? options.SpotlightUrl;
+
+        options.Experimental.EnableLogs = Experimental.EnableLogs ?? options.Experimental.EnableLogs;
 
 #if ANDROID
         Android.ApplyTo(options.Android);