+++ src/Sentry/Protocol/Envelopes/EnvelopeItem.cs
@@ -24,6 +24,7 @@ public sealed class EnvelopeItem : ISerializable, IDisposable
     internal const string TypeValueProfile = "profile";
     internal const string TypeValueMetric = "statsd";
     internal const string TypeValueCodeLocations = "metric_meta";
+    internal const string TypeValueLog = "log";
 
     private const string LengthKey = "length";
     private const string FileNameKey = "filename";
@@ -368,6 +369,21 @@ public sealed class EnvelopeItem : ISerializable, IDisposable
         };
 
         return new EnvelopeItem(header, new JsonSerializable(report));
+    }
+
+    [Experimental(Infrastructure.DiagnosticId.ExperimentalFeature)]
+    internal static EnvelopeItem FromLog(SentryLog log)
+    {
+        //TODO: allow batching Sentry logs
+        //see https://github.com/getsentry/sentry-dotnet/issues/4132
+        var header = new Dictionary<string, object?>(3, StringComparer.Ordinal)
+        {
+            [TypeKey] = TypeValueLog,
+            ["item_count"] = 1,
+            ["content_type"] = "application/vnd.sentry.items.log+json",
+        };
+
+        return new EnvelopeItem(header, new JsonSerializable(log));
     }
 
     private static async Task<Dictionary<string, object?>> DeserializeHeaderAsync(